requirements.txt
influxdb
hidapi
paho-mqtt
